+++ src/main/java/com/advella/advellabackend/services/ProductCategoryService.java
@@ -3,6 +3,7 @@ package com.advella.advellabackend.services;
 import com.advella.advellabackend.model.ProductCategory;
 import com.advella.advellabackend.repositories.IProductCategoryRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -18,15 +19,23 @@ public class ProductCategoryService {
         return productCategoryRepository.findAll();
     }
 
-    public void deleteProductCategory(int id) {
+    public ResponseEntity<Void> deleteProductCategory(int id) {
+        if (productCategoryRepository.getReferenceById(id) == null) {
+            return ResponseEntity.notFound().build();
+        }
         productCategoryRepository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 
     public void addNewProductCategory(ProductCategory productCategory) {
         productCategoryRepository.save(productCategory);
     }
 
-    public void updateNewProductCategory(ProductCategory productCategory) {
+    public ResponseEntity<Void> updateNewProductCategory(ProductCategory productCategory) {
+        if (productCategoryRepository.getReferenceById(productCategory.getProductCategoryId()) == null) {
+            return ResponseEntity.notFound().build();
+        }
         productCategoryRepository.save(productCategory);
+        return ResponseEntity.ok().build();
     }
 }
